step1b: multi-subject targets (per-subject mmap) + eval fix + disk-preflight sbatch
- targets_bigg: one memory-mapped file per subject (step1b_bigg_s{N}.pt) with a
  per-subject stats accumulator; pooled mean/std reconstructed by summation, so
  N-subject training never pulls 100GB+ of targets into RAM.
- data: keep targets in stored fp16 dtype, cast per-item (multi-subject RAM).
- eval_step1b: prefer the legacy single-file cache when it covers the requested
  subjects (subject-1 baseline reuses the 23GB cache, no re-encode); otherwise
  per-subject load. Fixes the broken torch.load of a never-written combined file.
- train sbatch: SUBJECTS env (default "1 2 5 7") + disk preflight that aborts
  before the build if free space < needed.
- add eval sbatch (slurm/eval_step1b_a100.sbatch).
- docs: correct cache size (~26GB/subject) + multi-subject instructions.

+++ brainflow/step1/data.py
@@ -20,7 +20,9 @@ class Step1Dataset(Dataset):
                  augment=False, noise_std=0.0):
         assert len(fmri) == len(target_emb) == len(images)
         self.fmri = ((fmri.float() - fmri_mu) / fmri_std)
-        self.emb = target_emb.float()
+        # Keep targets in their stored dtype (fp16, possibly memory-mapped) — do
+        # NOT .float() the whole array or multi-subject blows up RAM. Cast per item.
+        self.emb = target_emb
         self.images = images  # uint8
         self.subject_id = int(subject_id)
         self.augment = augment
@@ -36,7 +38,7 @@ class Step1Dataset(Dataset):
         img = self.images[i]
         if img.dtype == torch.uint8:
             img = img.float() / 255.0
-        return {"fmri": f, "emb": self.emb[i], "image": img, "subject": self.subject_id}
+        return {"fmri": f, "emb": self.emb[i].float(), "image": img, "subject": self.subject_id}
 
 
 def _collate(batch):
